+++ inst/doc/record_group.R
@@ -5,153 +5,151 @@ knitr::opts_chunk$set(
 )
 
 ## ----warning=FALSE, message=FALSE----------------------------------------
-library(diyar)
-library(dplyr)
-data(patient_list); patient_list
+library(diyar); library(dplyr)
+data(patient_list); 
+dbs <- patient_list[c("forename","surname","sex")]; dbs
 
-# Matching forename only
-cbind(patient_list, record_group(patient_list, rd_id, forename))
+# 1 stage <- Matching surname only
+dbs$pids_a <-record_group(dbs, criteria = surname, to_s4 = TRUE)
 
-# Matching forename and surname
-patient_list <- mutate(patient_list, cri_1 = paste(forename, surname,sep="-") )
-cbind(patient_list, record_group(patient_list, rd_id, cri_1, display = FALSE))
+# 2 stage - Matching surname, then matching sex
+dbs$pids_b <- record_group(dbs, criteria = c(surname, sex), display = FALSE, to_s4 = TRUE)
 
+dbs
 # Note that exact matching is case sensitive. See range matching.
 
 ## ----warning=FALSE, message=FALSE----------------------------------------
-cbind(patient_list, record_group(patient_list, rd_id, c(forename, surname), display = FALSE))
+dbs$pids_c <- record_group(dbs, criteria =  c(sex, surname), display = FALSE, to_s4 = TRUE)
+
+dbs
 
 ## ----message=FALSE, warning=FALSE----------------------------------------
-patient_list <- mutate(patient_list, cri_2 = paste(surname, sex,sep="-") )
-cbind(patient_list, record_group(patient_list, rd_id, c(forename, cri_2), display = FALSE))
+dbs_2 <- patient_list; dbs_2
+
+dbs_2$cri_2 <- paste(dbs_2$surname, dbs_2$sex,sep="-")
+dbs_2$pid_d <- record_group(dbs_2, rd_id, c(forename, cri_2), display = FALSE, to_s4 = TRUE)
+
+dbs_2
 
 ## ----message=FALSE, warning=FALSE----------------------------------------
 library(tidyr)
 data(Opes); Opes
 
 # 1 stage linkage
-  # stage 1 - name, and either department, hair colour or date of birth
-cbind(
-  Opes,
-  record_group(Opes, rd_id, name, list("s1a"=c("department","hair_colour","date_of_birth")),  display = FALSE)
-  ) %>% select(-starts_with("db_"), -sn)
+  # sub_criteria - name, and either department, hair colour or date of birth
+Opes$pids_a <- record_group(Opes, criteria = name, 
+                            sub_criteria = list("s1a"=c("department","hair_colour","date_of_birth")),
+                            display = FALSE, to_s4 = TRUE)
+
+Opes[c("name","department","hair_colour","date_of_birth","pids_a")]
   
 # 1 stage linkage 
-  # stage 1 - name, and either department or hair colour, and date of birth 
-cbind(
-  Opes,
-  record_group(Opes, rd_id, c(name), 
-               list("s1a"=c("department","hair_colour"), 
-                    "s1b"=c("date_of_birth")),  display = FALSE)
-  ) %>% select(-starts_with("db_"), -sn)
+  # sub_criteria - name, and either department or hair colour, and date of birth 
+Opes$pids_b <- record_group(Opes, criteria = name, 
+                            sub_criteria = list(
+                              "s1a"=c("department","hair_colour"),
+                              "s1b"=c("date_of_birth")),
+                            display = FALSE, to_s4 = TRUE)
+
+Opes[c("name","department","hair_colour","date_of_birth","pids_b")]
 
 # 1 stage linkage 
-  # stage 1 - name, and either department or hair colour, and either day and month of birth, day and year of birth or month and year of birth date of birth
-cbind(
-  Opes,
-  record_group(Opes, rd_id, c(name), 
-               list("s1a"=c("department","hair_colour"), 
-                    "s1b"=c("db_pt1","db_pt2","db_pt3")),  display = FALSE)
-  ) %>% select(-date_of_birth, -sn)
+  # sub_criteria - name, and either department or hair colour, and either day and month of birth, day and year of birth or month and year of birth
+Opes$pids_c <- record_group(Opes, criteria = name, 
+                            sub_criteria = list(
+                              "s1a"=c("department","hair_colour"), 
+                              "s1b"=c("db_pt1","db_pt2","db_pt3")),
+                            display = FALSE, to_s4 =TRUE)
 
+Opes[c("name","department","hair_colour","date_of_birth","pids_c")]
 
 # 1 stage linkage 
-  # stage 1 - name, and department, and hair colour, and either day and month of birth, day and year of birth or month and year of birth date of birth 
-cbind(
-  Opes,
-  record_group(Opes, rd_id, c(name), 
-               list("s1a"=c("department"),
-                    "s1c"=c("hair_colour"),
-                    "s1b"=c("db_pt1","db_pt2","db_pt3")),  display = FALSE)
-  ) %>% select(-starts_with("db_"), -sn)
+  # sub_criteria - name, and department, and hair colour, and either day and month of birth, day and year of birth or month and year of birth
+Opes$pids_d <- record_group(Opes, criteria =name, 
+               sub_criteria = list(
+                 "s1a"=c("department"),
+                 "s1c"=c("hair_colour"),
+                 "s1b"=c("db_pt1","db_pt2","db_pt3")),  
+               display = FALSE, to_s4 = TRUE)
 
+Opes[c("name","department","hair_colour","date_of_birth","pids_d")]
 
 ## ---- warning=FALSE, message=FALSE---------------------------------------
 # 1 stage linkage 
   # stage 1 - name, and date of birth, and department and hair colour 
-cbind(
-  Opes,
-  record_group(Opes, rd_id, name, 
-               list("s1a"=c("department"),
-                    "s1b"=c("hair_colour"),
-                    "s1c"=c("date_of_birth")),  display = TRUE)
-  ) %>% select(-starts_with("db_"))
+Opes$pids_e <- record_group(Opes, criteri = name, 
+                            sub_criteria = list(
+                              "s1a"=c("department"), 
+                              "s1b"=c("hair_colour"), 
+                              "s1c"=c("date_of_birth")),
+                            display = TRUE, to_s4 = TRUE)
 
-# 1 stage linkage 
-  # stage 1 - name, and date of birth, and department and hair colour
-Opes_b <- unite(Opes, cri, c(name, date_of_birth, department, hair_colour))
-cbind(
-  Opes_b,
-  record_group(Opes_b, rd_id, c(cri),  display = TRUE)
-  ) %>% select(-starts_with("db_"))
+Opes$cri <- paste(Opes$name, Opes$date_of_birth, Opes$department, Opes$hair_colour, sep="-")
+
+Opes$pids_f <- record_group(Opes, criteria = cri,  display = TRUE, to_s4 =TRUE)
+
+Opes[c("name","department","hair_colour","date_of_birth","pids_e","pids_f")]
 
 ## ----message=FALSE, warning=FALSE----------------------------------------
 library(lubridate)
-Opes_c <- select(Opes, date_of_birth)
-Opes_c$dummy_cri <- 1
+Opes_c <- Opes["date_of_birth"]
 Opes_c
 
-# Match record within 3 months before or after a date
-Opes_c$range <- expand_number_line(as.number_line(dmy(Opes_c$date_of_birth)), period(2, "years"), "end")
-Opes_c$range@gid <- as.numeric(dmy(Opes_c$date_of_birth))
-bind_cols(Opes_c,
-  record_group(Opes_c, criteria = dummy_cri, sub_criteria = list(s1="range")))
+# Match on date of birth + 3 months
+Opes_c$date_of_birth <- dmy(Opes_c$date_of_birth)
+Opes_c$range_a <- expand_number_line(as.number_line(Opes_c$date_of_birth), period(2, "years"), "end")
+Opes_c$range_a@gid <- as.numeric(Opes_c$date_of_birth)
 
-# Match record within 5 years younger or older than an age
-Opes_c$age <- as.numeric(round((Sys.Date() - dmy(Opes_c$date_of_birth))/365.5)) # approximate age
-Opes_c$range <- as.number_line(Opes_c$age)
-Opes_c$range@gid <- Opes_c$age
-Opes_c$range <- expand_number_line(Opes_c$range, 5, "end")
-bind_cols(Opes_c,
-  record_group(Opes_c, criteria = dummy_cri, sub_criteria = list(s1="range")))
+Opes_c$pids_a <- record_group(Opes_c, criteria = range_a, to_s4 =TRUE)
+
+Opes_c[c("date_of_birth","range_a","pids_a")]
+
+# Match on age +/- 5 years
+Opes_c$age <- as.numeric(round((Sys.Date() - Opes_c$date_of_birth)/365.5)) # approximate age
+Opes_c$range_b <- expand_number_line(as.number_line(Opes_c$age), 5, "both")
+Opes_c$range_b@gid <- Opes_c$age
+
+Opes_c$pids_b <- record_group(Opes_c, criteria = range_b, to_s4 =TRUE)
+
+Opes_c[c("age","range_b","pids_b")]
 
 ## ----warning=FALSE, message=FALSE----------------------------------------
 data(patient_list_2); patient_list_2
 
-cbind(
-  patient_list,
-  record_group(patient_list, rd_id, c(forename, surname, sex))
-)
+patient_list_2$pids_a <- record_group(patient_list_2, rd_id, c(forename, surname, sex), to_s4 = TRUE)
+
+patient_list_2
 
 ## ----warning=FALSE, message=FALSE----------------------------------------
-patient_list_b <- patient_list_2
-patient_list_b <- mutate(patient_list_b, forename = 
-                           ifelse(rd_id %in% 1:3, "Nil", forename))  
-
+patient_list_2$forename <- ifelse(patient_list_2$rd_id %in% 1:3, "Nil", patient_list_2$forename)
 # 2 stage linkage
     # Stage 1 - forename
     # Stage 2 - Surname
-cbind(
-  patient_list_b,
-  record_group(patient_list_b, rd_id, c(forename, surname), display = FALSE)
-)
+patient_list_2$pids_b <- record_group(patient_list_2, criteria = c(forename, surname), 
+                                      display = FALSE, to_s4 =TRUE)
 
 # 2 stage linkage
     # Stage 1 - forename
     # Stage 2 - Surname and sex
-patient_list_b <- mutate(patient_list_b, cri_2 = paste(surname,sex,sep=""))
+patient_list_2 <- mutate(patient_list_2, cri_3 = paste(surname,sex,sep=""))
 
-cbind(
-  patient_list_b,
-  record_group(patient_list_b, rd_id, c(forename, cri_2), display = FALSE)
-)
+patient_list_2$pids_c <- record_group(patient_list_2, criteria = c(forename, cri_3), 
+                                      display = FALSE, to_s4 = TRUE)
 
+patient_list_2[c("forename","surname","sex","pids_b","pids_c")]
 
 ## ----warning=FALSE, message=FALSE----------------------------------------
 # Using NA as the proxy for missing value
-patient_list_b <- mutate(patient_list_b,forename = ifelse(forename=="Nil",NA,forename))
+patient_list_2 <- mutate(patient_list_2,forename = ifelse(forename=="Nil",NA,forename))
 
-cbind(
-  patient_list_b,
-  record_group(patient_list_b, rd_id, c(forename, surname), display = FALSE)
-)
+patient_list_2$pids_d <- record_group(patient_list_2, rd_id, c(forename, surname), 
+                                      display = FALSE, to_s4 = TRUE)
 
 # Using "" as the proxy for missing value
-patient_list_b <- mutate(patient_list_b,forename = ifelse(is.na(forename),"",forename))  
+patient_list_2 <- mutate(patient_list_2,forename = ifelse(is.na(forename),"",forename))  
 
-cbind(
-  patient_list_b,
-  record_group(patient_list_b, rd_id, c(forename, surname), display = FALSE)
-)
+patient_list_2$pids_e <- record_group(patient_list_2, rd_id, c(forename, surname), 
+                                      display = FALSE, to_s4 = TRUE)
 
+patient_list_2[c("forename","surname","sex","pids_d","pids_e")]